Move member to simplelayout folder
