+++ ftw/contacts/simplelayout/member.py
